Rewrite effects/colide.js
The code in colide.js was disorganized.

To organize a new folder was created `effects/colide`.
The sylvester module and `sylvester.js` was moved to this new folder.

Part of the code of colide.js was moved to `effects/colide/physics.js`.
The old code was rewritten, and the part moved to `physics.js`
was refactored.

To accomodate this refactor jstools was modified to support
overwritting of priviledged functions.

TODO: colide.js new functions were not tested yet.