Add working version of colide.js & minor bug fixing
This commit finished the first working version of colide.js,
it also adds a big a test batery implemented with mocha.js,
to test `colide.js`, the tests are at `effects/test` folder.

The exaustive testing suit built to test colide.js, also found
bugs on body.js and sugested a refactor on space.js, thats why
they are also modified in this commit.

The Colision class was moved to a separated file for its importance,
and was improved to be more reliable and to have some useful functions.